Feature/user endpoint

+++ src/main/java/com/booking/project/exception/ResourceConflictException.java
@@ -3,18 +3,18 @@ package com.booking.project.exception;
 public class ResourceConflictException extends RuntimeException{
     private static final long serialVersionUID = 1791564636123821405L;
 
-    private int resourceId;
+    private Long resourceId;
 
-    public ResourceConflictException(int resourceId, String message) {
+    public ResourceConflictException(Long resourceId, String message) {
         super(message);
         this.setResourceId(resourceId);
     }
 
-    public int getResourceId() {
+    public Long getResourceId() {
         return resourceId;
     }
 
-    public void setResourceId(int resourceId) {
+    public void setResourceId(Long resourceId) {
         this.resourceId = resourceId;
     }
 }
